Add NonRegisteringDriver to allow the use of the driver without registering with the DriverManager
