+++ src/main/java/org/threadly/db/aurora/Driver.java
@@ -1,14 +1,7 @@
 package org.threadly.db.aurora;
 
-import java.sql.Connection;
 import java.sql.DriverManager;
-import java.sql.DriverPropertyInfo;
 import java.sql.SQLException;
-import java.sql.SQLFeatureNotSupportedException;
-import java.util.Properties;
-import java.util.logging.Logger;
-
-import org.threadly.db.AbstractArcDriver;
 
 /**
  * Threadly's AuroraArc Driver.  This Driver will create multiple connections for each returned 
@@ -28,51 +21,22 @@ import org.threadly.db.AbstractArcDriver;
  * <li>{@code "optimizedStateUpdates=true"} - Experimental internal code that can provide performance gains
  * </ul>
  */
-public class Driver extends AbstractArcDriver {
+public class Driver extends NonRegisteringDriver implements java.sql.Driver {
   static {
     try {
       DriverManager.registerDriver(new Driver());
     } catch (SQLException e) {
-      throw new RuntimeException(e);
+      throw new RuntimeException("Can't register driver!", e);
     }
   }
   
   /**
-   * Another way to register the driver.  This is more convenient than `Class.forName(String)` as 
-   * no exceptions need to be handled (instead just relying on the compile time dependency).
+   * Construct a new driver and register it with DriverManager
+   * 
+   * @throws SQLException
+   *             if a database error occurs.
    */
-  public static void registerDriver() {
-    // Nothing needed, just a nicer way to initialize the static registration compared to Class.forName.
-  }
-
-  @Override
-  public Connection connect(String url, Properties info) throws SQLException {
-    if (DelegatingAuroraConnection.acceptsURL(url)) {
-      return new DelegatingAuroraConnection(url, info);
-    } else {
-      // JDBC spec specifies that if unhandled URL type at this point is provided, null should be returned
-      return null;
-    }
-  }
-
-  @Override
-  public boolean acceptsURL(String url) {
-    return DelegatingAuroraConnection.acceptsURL(url);
-  }
-
-  @Override
-  public DriverPropertyInfo[] getPropertyInfo(String url, Properties info) throws SQLException {
-    return DelegateDriver.getDriver().getPropertyInfo(url, info);
-  }
-
-  @Override
-  public boolean jdbcCompliant() {
-    // should be compliant since just depending on mysql connector
-    return DelegateDriver.getDriver().jdbcCompliant();
-  }
-
-  @Override
-  public Logger getParentLogger() throws SQLFeatureNotSupportedException {
-    return DelegateDriver.getDriver().getParentLogger();
+  public Driver() throws SQLException {
+      // Required for Class.forName().newInstance()
   }
 }
